+++ src/main/kotlin/com/yanncebron/intellipikchr/editor/PikchrEditorWithPreview.kt
@@ -16,8 +16,14 @@
 
 package com.yanncebron.intellipikchr.editor
 
+import com.intellij.openapi.actionSystem.ActionGroup
+import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.fileEditor.TextEditor
 import com.intellij.openapi.fileEditor.TextEditorWithPreview
 
 class PikchrEditorWithPreview(editor: TextEditor, preview: PikchrPreviewFileEditor) :
-    TextEditorWithPreview(editor, preview, "Pikchr Split Editor", Layout.SHOW_EDITOR_AND_PREVIEW)
+    TextEditorWithPreview(editor, preview, "Pikchr Split Editor", Layout.SHOW_EDITOR_AND_PREVIEW) {
+    override fun createRightToolbarActionGroup(): ActionGroup? {
+        return ActionManager.getInstance().getAction("PikchrEditorRightToolbarGroup") as ActionGroup?
+    }
+}